// src/generated/models/Session.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Thread } from './Thread';
export type Session = {
    id?: string;
    thread?: Thread;
    prompt?: string;
    response?: string;
    context?: string;
    createdAt?: string;
};