// src/generated/models/Thread.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Session } from './Session';
export type Thread = {
    id?: string;
    paperId?: string;
    title?: string;
    context?: string;
    sessions?: Array<Session>;
    createdAt?: string;
    updatedAt?: string;
};